chore: add api serve stop subcommand - update -h to -a
